# openapi.json
{
   "openapi": "3.0.0",
   "info": {
      "title": "Task API",
      "version": "1.0.0",
      "description": "A minimal CRUD API for managing tasks, built with Express."
   },
   "servers": [
      { "url": "http://localhost:3000" }
   ],
   "paths": {
      "/tasks": {
         "get": {
         "summary": "List all tasks",
         "responses": {
            "200": {
               "description": "Array of tasks",
               "content": {
               "application/json": {
                  "schema": {
                     "type": "array",
                     "items": { "$ref": "#/components/schemas/Task" }
                  }
               }
               }
            }
         }
         },
         "post": {
         "summary": "Create a new task",
         "requestBody": {
            "required": true,
            "content": {
               "application/json": {
               "schema": {
                  "type": "object",
                  "properties": {
                     "title": { "type": "string" }
                  },
                  "required": ["title"]
               }
               }
            }
         },
         "responses": {
            "201": {
               "description": "Task created",
               "content": {
               "application/json": {
                  "schema": { "$ref": "#/components/schemas/Task" }
               }
               }
            },
            "400": { "description": "Missing or empty title" }
         }
         }
      },
      "/tasks/{id}": {
         "get": {
         "summary": "Get a single task",
         "parameters": [
            { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
         ],
         "responses": {
            "200": {
               "description": "The requested task",
               "content": {
               "application/json": {
                  "schema": { "$ref": "#/components/schemas/Task" }
               }
               }
            },
            "404": { "description": "Task not found" }
         }
         },
         "put": {
         "summary": "Update a task",
         "parameters": [
            { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
         ],
         "requestBody": {
            "required": true,
            "content": {
               "application/json": {
               "schema": {
                  "type": "object",
                  "properties": {
                     "title": { "type": "string" },
                     "done": { "type": "boolean" }
                  }
               }
               }
            }
         },
         "responses": {
            "200": {
               "description": "Task updated",
               "content": {
               "application/json": {
                  "schema": { "$ref": "#/components/schemas/Task" }
               }
               }
            },
            "400": { "description": "Invalid body" },
            "404": { "description": "Task not found" }
         }
         },
         "delete": {
         "summary": "Delete a task",
         "parameters": [
            { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
         ],
         "responses": {
            "204": { "description": "Task deleted" },
            "404": { "description": "Task not found" }
         }
         }
      }
   },
   "components": {
      "schemas": {
         "Task": {
         "type": "object",
         "properties": {
            "id": { "type": "integer" },
            "title": { "type": "string" },
            "done": { "type": "boolean" }
         }
         }
      }
   }
}